// src/app/shared/models/book-genre.enum.ts
export enum BookGenre {
  ActionAdventure = 'Action & Adventure',
  Art = 'Art',
  Autobiography = 'Autobiography',
  Biography = 'Biography',
  Business = 'Business',
  Children = 'Children',
  Classic = 'Classic',
  ComicsGraphicNovel = 'Comics & Graphic Novel',
  Contemporary = 'Contemporary',
  Crime = 'Crime',
  Drama = 'Drama',
  Dystopian = 'Dystopian',
  Education = 'Education',
  Essay = 'Essay',
  Fantasy = 'Fantasy',
  HistoricalFiction = 'Historical Fiction',
  History = 'History',
  Horror = 'Horror',
  LiteraryFiction = 'Literary Fiction',
  Memoir = 'Memoir',
  Mystery = 'Mystery',
  Philosophy = 'Philosophy',
  Poetry = 'Poetry',
  Politics = 'Politics',
  Programming = 'Programming',
  Psychology = 'Psychology',
  ReligionSpirituality = 'Religion & Spirituality',
  Romance = 'Romance',
  Science = 'Science',
  ScienceFiction = 'Science Fiction',
  SelfHelp = 'Self-Help',
  ShortStories = 'Short Stories',
  Sociology = 'Sociology',
  Technology = 'Technology',
  Thriller = 'Thriller',
  Travel = 'Travel',
  War = 'War',
  YoungAdult = 'Young Adult',
}

export const BOOK_GENRE_OPTIONS: Array<{ label: string; value: BookGenre }> = Object.values(
  BookGenre,
).map(v => ({ label: v, value: v as BookGenre }));